trunk: [signal_handler] corrected integer kind

+++ src/utils/signal_handler.F90
@@ -58,7 +58,7 @@ contains
       use ifport, only: signal
 #endif /* __INTEL_COMPILER */
       implicit none
-      integer, intent(in) :: signum
+      integer(kind=4), intent(in) :: signum
       procedure(signal_handler) :: func
       integer :: err
